Fixed some variables shared between threads
I removed some Arc<Mutex<_>> because cellagent.rs has only one thread
talking to the cmodel.  Bill pointed out that there is also a thread
talking to the up tree.  Oopsies!  The four cell agent variables used
by both are traph, saved_msgs, saved_stack, and tree_name_map.  They
are all now wrapped in Arc<Mutex<_>>.

Rust does not compile if there is a data race, but I am cloning the
cell agent for each thread, so each one is working on its own copy of
the data structures.  The mystery is why updates in one thread show up
in the other at all when the variables aren’t wrapped.  It must be
related to the problem I had with the failover race when working with a
clone.