Fixed missing entry in TreeIDMap
A cell gets a new port tree id on receiving a FailoverDMsg, but I
wasn’t adding it to the TreeIDMap.  The result was a panic in debug
print in cell agent update_element.

I discovered the Either crate, which lets me handle saved_msgs in a
more principled way.

I also cleaned up a bunch of warning messages.