Use ThreadLocal instead of passing trace_header
Having to pass trace_header through several layers of calls has annoyed
me since the day I did it.  Bill suggested using ThreadLocal so that
functions that needed the trace header can get it from the thread
global state.  I tried to get it to work a couple of weeks ago, but the
event ids were wrong.  I figured out my mistake, and this version
appears to give essentially the same trace records.

I noticed that there were two places where the trace header was forked
when it didn’t need to be, so I removed them.  Aside from that change,
the trace records from this version should be equivalent to those from
before the change.