Port tree forwarding table entries correct before breaking links
On the root of a tree each port tree has exactly one child.  The
corresponding tree has a mask that is the OR of all of them.  On a cell
other than the root of the tree, the port tree that points to the
parent has all the children of the tree, and the others are dummy until
a failover occurs.