Fixed a bug in recording ports already tried during a failover
I was inconsistent in whether I used tree_id or port_tree_id when
recording which ports had already been tried for a failover.  I am now
consistently using port tree id so that I can deal with multiple link
failures from the same cell.

I also updated Display for TreeID so that it prints the first 8
characters of the UUID in addition to the name.  That lets me easily
distinguish a tree id from a port tree id.  It also helps debugging the
packet engine, which only deals with UUIDs.