Set forwarding table entries for tree, port_trees, and stacked trees in root ward cell
The Failover messages get through to the root ward cell, but there’s
something wrong when the second one goes through the same port as the
first.  Even so, the forwarding table entries and traph state to
reflect the failover path are correct for the broken tree, port trees,
and stacked trees at the root ward cell of the broken link.

In the process of debugging, I moved updates to the traph data
structures into traph.rs, which I use in process_failover_msg.  I’ll be
changing updates in other places to use this new code as I find it.

I also added a starting_epoch to TraceHeader to help Gregory figure out
which Kafka records belong to the same run.