Applied most of clippy recommended changes
Clippy is a linter for Rust that used to be available only on nightly.
In Rust 2018 it is available on stable.  It gave me a lot of
recommendations to make the code conform to the standard style.

In several places, it recommended changes that improve performance.
For example, an argument with a type of &Vec<_> requires a new
allocation, while &[_] does not.

I did not accept some of the recommendations because I think the code
is clearer the way I had done them.