Rust just saved me from a big boo boo
In cellagent.rs, self.traphs is wrapped in a mutex because it’s read in
one thread and written in another.  I had a get_traphs method that
unlocked the mutex, got the traph, and returned a clone.  That’s fine
for read-only use of the traphs but not for updating it.

In order to change the traph, I was removing the traph from the
unlocked self.traphs, working on it, and putting it back.  The problem
is that the other thread could have inserted a traph with the same
tree_id.  RACE CONDITION.

I found taking the traph out and putting it back annoying, so I did
some investigation and found that I needed to pass the locked mutex
into a get_traph_mut method in order to return a mutable reference.
That means I’m holding the lock until I finish the update.  Voila, no
more race condition.