Got my port trees and trees straight
I hope.  There were three separate errors.  First, the DiscoverMsg
always had a port tree with port 0 instead of the actual root port.
That error hid two other errors.

The first one was failing to update the base tree map for port trees
other than the first one seen.  The second error was using a port tree
for quenching.  That worked as long as the port number in the
port_tree_id was zero, but it never quenched when it wasn’t, resulting
in Discover messages propagating forever.