A Rustier way to implement fmt::Display
I was doing a lot of s = s + format!(…), which has two downsides.  The
one that doesn’t matter is that each s + allocates a new string.  The
one that matters marginally, is that each macro occurrence gets
expanded at compile time.  The result is a larger executable and a
longer compile time.  The reason the benefit is marginal is that
neither of these is a problem now.  Still, it’s better to do  it right.