Traph now shows on console output
I fixed the problem of traphs not showing up on the console by created
a copy that is wrapped in Arc<Mutex<_>>.  This copy is only to get
around the problem I had printing a cell’s traphs from main.rs.  Hence,
I no longer need to put the traphs into the trace.