Added root_port_no to DiscoverMsg and DiscoverDMsg
John’s display with root port info is sometimes wrong.  The problem is
that Gregory is using Discover messages and doesn’t know which one to
use.  It’s not a problem for simple quenching, but it is for root port
quenching.

Gregory originally wanted to use DiscoverD messages for the translation
to John’s API, but he couldn’t because it didn’t have the root port
information.  This version adds the missing info.

This version includes some incomplete changes needed for the routing
table changes during failover.