Fixed a problem with console output
Since the cell agent runs in a thread, the console can only access
items wrapped in Arc<Mutex<_>>, but ca.traphs isn’t any more.  I got
around the problem by creating a shadow of ca.traphs that is.  The
problem is that I have to remember to update the shadow every time I
change a traph.  This commit fixes a place where I hadn’t.