Fixed problem with console output
It’s nice to be able to ask to see the state of a cell on the console
from main.  I introduced a shadow item traphs_mutex of type
Arc<Mutex<Vec<Traph>>> so I could do that, but I wasn’t updating it
every time I updated a traph.  What I do now is clone self.traphs into
self.traphs_mutex at the end of every method that uses
self.get_traphs_mut().  That’s probably so slow I’ll have to pull it
out eventually, but it’s convenient for now.