Fixed problem with finding trees with broken links
When a link breaks, each of the cells it connects labels its
corresponding port as broken in each traph element.  I was doing that
in a map() in a statement with find().  The problem is find() stops
asking map() for more elements once find() gets a true.  The result was
that not all traph elements had the port marked broken.  To make
matters worse, the hash map keys are UUIDs, which change on every run.
That meant a different set of elements didn’t get set, which made it
look like a race condition.

The fix is to move setting the ports as broken into a for loop.