Made Name Copy
I changed the name field of Name from String to [char;32], which
allowed me to make all the IDs (e.g., CellID, TreeID) Copy.  The result
was that I could remove .clone() from a gazillion places in the code.

Right now the name field is limited to 32 characters.  The limit can be
extended with an explicit impl for each ID.