Added ENTL type in uuid.rs and save last packet sent in packet_engine.rs
In order to sew up the packet stream following a failover, leaf ward
needs to resend the last packet in case root ward  didn’t get it.  The
packet engine keeps that packet in a HashMap<PortNo,Packet>.  When the
sending cell receives a packet on that port, it removes the entry.

If root ward did get it and has no data packets to send, it sends an
ENTL packet to leaf ward, so leaf ward knows it can remove the entry
for that port.