Fixed Kafka problem
I was using a domain name that I put into /etc/hosts for the URL of the
Kafka server.  That worked for a long time, but something changed.  I
replaced the domain name with the explicit IP address in config.rs,
which means I’ll have to manually update it when the server gets a new
address.

I also removed the ENTL packet code, since TICK is an ENTL packet that
does what I want.  In the process, I changed the AIT codes so their lsb
is always 1.  That makes them easy to identify in the debugger, since
their values are always odd numbers.