Setting up data structures for flow control
and, as it turns out, sewing the packet stream back up on failover.
The basic idea is like ENTL; the sender of a packet blocks until the
receiver sends an event, either a packet or an ENTL.  The receiver
doesn’t send that event until it can be delivered the packet to the
outgoing ports.

This form of credit based flow control can be used to avoid dropping
packets (except in the case of node failure).  All that’s needed is for
the sender to keep a copy of the last packet sent until the
acknowledging event arrives.  If the sender remembers how many packets
it sent between receive events, and the receiver remembers how many
packets received, it’s easy to sew the packet stream back up with no
lost or duplicated packets.