Conditionally forward packets
Until now, there’s been no flow control.  This commit has the first
part of an implementation.  If the out port is in the can_send state,
it will forward up to N packets.  N is the number of packets that can
be sent before needing an event from the other side of the link.  N is
encoded in config.rs as PACKET_PIPELINE_SIZE.

There is a commented line in send_packet that sets can_send to false.
I can’t uncomment that until I’ve added the ability to set can_send to
true when receiving a packet.