Flow control works when I turn it off
That doesn’t sound like much, but the code had been locking up even
when flow control was off.  The problem was the handling of struct
items that were used by both the cm and port loops.  They run in
separate threads and were working off different clones.  The result was
that setting a value in one thread had no effect on the other.  The
solution was to wrap them in Arc<Mutex<_>>.