Fixed bad trace problem
I had made some changes to the TRACE_OPTIONS and DEBUG_OPTIONS that
resulted in missing trace records.  The problem was a confusion between
DEBUG_OPTIONS.all and TRACE_OPTIONS.all.  I went through all the uses
and cleaned up the code so it is easier to read.  My hope is that will
make it easier for me to avoid making this kind of mistake again.