Setting up for sewing up the packet stream
I added a NumberOfPackets struct to packet_engine.rs, which holds the
number of sent and seen packets.  I’ve added that field to the Pe->Cm
and Cm->Ca paths so the cell agent gets these numbers when a link
breaks.  I also added a Failover to Ca->Cm and Cm->Pe to contain
whatever data the packet engine will need following a successful
failover.

I also put all the trace blocks in an empty block as I had done in the
commit just before the last one.  It’s easier for my eye to skip it,
and it’s easier to tell the IDE to collapse the block.