Added NumberOfPackets to FailoverMsg and FailoverDMsg
In order to sew up the packet stream following a successful failover,
leaf ward needs to know if it needs to resend the last packet sent
before the link broke.  To that end, each packet engine keeps a count
of packets sent and received modulo the pipeline depth.

Leaf ward puts its value in the FailoverMsg and root ward puts that and
its value in the FailoverDMsg.  When leaf ward receives the
FailoveDMsg, it can tell the packet engine whether or not it needs to
resend the last packet.

This version does all that except informing the packet engine.