Added field to Packet struct for failover wrapping
When having RW sew up the packet stream, LW needs to wrap packets being
rerouted.  Since the new route may also cross a broken link, I need a
stack of wrapped tree uuids.  I’ve chosen to do that with a Vec, which
is not Copy.  This commit makes the necessary changes.