Added wrap and unwrap methods to packet.rs
When sewing up the packet stream on root ward, we wrap the packet that
would have gone out on the broken port in a header that points to
rw_tree_id.  The root cell will unwrap that packet to restore the
original header and send the packet on its way.