Reroute packet following failover
The code to reroute packets is in place, but building a test case will
take a non-trivial amount of work.

The hard part is keeping packet order.  There are several options, but
I’ve chosen to keep the forwarding table entries pointing to the port
connected to the broken link.  The packet engine takes care of moving
those packets to the failover port.