Added a , to the end of each record written to the trace file
JSON editors are really picky.  If your JSON isn’t perfect, most of
them just say, “You have an error.”  Adding a , to the end of each
record makes it easier to manually fix up the file.  Just add a [ at
the start and replace the , at the end with a ].  Voila!  You’ve got
well-formed JSON.

I could put in the [ automatically, but I don’t know when I’ve put the
last record into the file, which means I can’t put  in the
corresponding close the bracket.