Added RACE_SLEEP to config.rs to hide race condition
It will take me a while to fix the race condition, so I added
RACE_SLEEP to config.rs that let’s us hide it.  What it does in noc.rs
is sleep RACE_SLEEP seconds to let the discover process finish and
RACE_SLEEP seconds to let tree stacking finish.  A value of 4 seconds
seems to guarantee that every cell gets all its messages; a value of 2
seconds succeeds most of the time.