Removed last vestige of blocking messages
In order to fix a message race, I introduced the idea of messages that
block the packet engine until they are processed.  Not only is that bad
for performance, it only made the message race less frequent.  The
approach also forced the cell agent to hold messages to be forwarded
later, another bad idea.

Rather than trying to debug the message race with the design I have
now, I’m completely changing the way I deal with the problem.  The
first step is to remove all code dealing with blocking.