Getting ready to send DiscoverD NACK
In the current version, a DiscoverD is sent only the first time a cell
sees a given tree ID.  That means a cell never knows if it’s heard if
its other neighbors have heard of that tree ID.

This commit prepares for sending two kinds of DiscoverD, First and
Subsequent.  However, this version never sends Subsequent.