Added DiscoverD NACK
Before this commit ports were initialized in the Pruned state.  On
receiving a DiscoverMsg for the first time for a given tree, a
DiscoverDMsg was sent.  The sender marked the link Parent, and the
recipient marked its in-port Child.  No message was sent on subsequent
receipt of a Discover message on that tree.  The result was that cells
didn’t know if a Pruned link might be turned into a Child.

I am now sending a DiscoverDMsg::First on first receiving a DiscoverMsg
for a tree and a DiscoverDMsg::Subsequent on receiving additional
DiscoverMsgs for that tree.  Ports start in an Unknown state and get
turned into Pruned when receiving a DiscoverDMsg::Subsequent.

I also moved all explicit module names to use statements because of a
glitch in the IntelliJ editor.  It reports errors saying that some
crates are private even though they are not, as shown by the
application compiling and running.  Moving the explicit module names
reduces the number of spurious red marks in the body of the code.  It’s
still annoying, but the annoyance is more manageable.