Removed saved application msgs from cell agent
When pipelining messages, a new child can be added to a stacked tree
after an application message has arrived.  That message may need to be
saved so that it can be forwarded to the new child.  I had saved these
application messages in the cell agent, which is a mistake; the
application should be responsible for saving those messages.  This
commit removes saved messages from the cell agent but does not provide
for saving them in the application.  Hence, they simply do not get
forwarded to new children.

I also removed the last vestiges of blocking the packet engine to
prevent messages from racing ahead when the cell agent is updating a
tree.  The new way of handling application messages makes this strategy
unnecessary.