Moved AIT handling from packet engine to port
I had been doing the two AIT round trips packet engine to packet
engine, but I should have been doing them port to port.  That’s what
this version does.

There is a minor complication due to the way I do flow control.  The
packet engine needs to know when an ENTL packet is received, so it
knows it’s its turn to send a data packet.  I may not need that on the
physical system.

This version, like the pe to pe version, does not inform the sender
that the transfer succeeded or failed.  There’s a TODO in the code, so
I don’t forget to add it once I figure out what to do when the transfer
fails.

I spent several days tracking down a bug that wasn’t a bug.  Because I
haven’t put code in place to retry application messages when a late
child joins, I sleep for a few seconds to let the tree stacking finish
before sending those messages.  However, in order to debug the AIT
process, I turned on trace records for the port.  There were so many of
them that I wasn’t waiting nearly long enough.  Without the port trace
records, I only need to wait 4 seconds; with them, 30 seconds.