Actually use AppMessage
Previously, I was using an ad hoc way to get strings from the Noc and
applications to the cell agent.  This version actually uses the
AppMessage structs.

It also uses the typetag crate for the app messages.  typetag lets
serde deserialize a trait object without telling it the explicit type.
I’ll do that for EC messages in a later commit.

The code runs but stops before deploying NocMaster and NocAgent.  I’m
doing this commit because it includes so many change