Encapsulated ByteArray
ByteArray has a Vec<u8>, which was exposed to all users of it.  In
other words, they did string-u8 conversions.  That wasn’t a good idea.
All those conversions should be done in the impl for ByteArray.

One reason I did the change now is that messages aren’t getting
through, and it’s really hard to figure out what I’m seeing in the
trace records for packet_engine and port.  With this change I should be
able to switch between bytes and strings more easily.