Use typetag for ec_messages
When I receive an EC message of trait Message, the serde deserializer
needs to know the explicit message type.  I had been doing a double
serialization, serialize the message then serialize the message type
and the serialized message together.  On deserialization, I’d
deserialize to get the type and use that type to deserialize the actual
message.  Yuck.

The typetag crate does that for me in a more principled way.