Fixed trace records in packet_engine.rs
I had a whole bunch of sends without trace records.  I was originally
going to put the sends in functions so I could call them from multiple
places.  I decided not to do that because it will help if I know what
loop I was in when I did the send.  I could still have used a function
by passing in the function name and appropriate line number, but I
decided to do the simple thing first.