Fixed handling of inter-application messages
An EC inter-application message carries the App inter-application
message in the payload.  All I needed to do was extract it, serialize
it, and send it on its way.

I could probably put the serialized app message in the payload, but
this approach works even if the serialization is different on the
sending and receiving sides.