Fixed a problem in handling manifest message
In app_manifest() in cellagent.rs, I was extracting the tree_map from
self.tree_name_map but not holding the lock.  That meant the tree_map
could be changed while I was updating it.  Then, when I stored the
changes, I’d overwrite the other ones.

This version holds the lock until the updates are made.