Replaced a hashmap.remove() with hash map.entry()
In order to avoid borrow problems in two places in cellagent.rs, I was
removing an item from a hash map, modifying it, and putting it back.
The problem is that the item isn’t put back on an error return.

Rust has a way to update a hashmap item in-place called
hashmap.entry().  I was able to use that in one place.  The other still
runs into a borrow problem, so I clone the value and explicitly insert
it after updating.