Stack tree quiescence implemented
Assume that the discover process has quiesced, which means no more
cells will join the black tree.  I’ll deal with adding cells to the
datacenter later.

I had been doing tree stacking in such a way that new cells could join
the stacked tree while messages were in flight.  In that scheme, a cell
sent a StackTreeD immediately on receiving the StackTree message.  The
problem was dealing with messages in flight when a new cell joined the
stacked tree.

I avoid that problem by only sending a StackTreeD when a cell has
received one from each of its children on the parent tree.  The
recursion ends when a leaf on the parent tree receives a StackTree
message.  With this scheme there is no way to send a message on a tree
before it is built out.

Cells have to send a StackTreeD NACK if they are not joining the tree.
They can also end the recursion if the GVM XTND equation evaluates to
false.