Change deploy process to be compatible with new tree building
When I was building trees from root to leaf, I could just send a deploy
message immediately after sending the stack tree message.  However, I
had to deal with late child arrivals.  I now build the trees from leaf
to root, which means I can’t send a message on a tree until it is
completely built (assuming that the Discover process has quiesced).

This version also properly sets up and verifies the AllowedTree to
TreeID c-list for each sender, something that I had been doing
inconsistently.  In the process of making this change, I verified that
the approach does indeed prevent using a tree not in the c-list.