Preparing to build base tree from leaf to root
I currently build the base tree from root to leaf.  The problem is that
the root doesn’t know when the tree building has quiesced.  Instead, I
will build the tree from leaf to root.  Now, when the root cell gets
the last DiscoverD, it knows the tree is complete.

I need a heuristic for a cell to know when it has gotten a DiscoverD
for a particular tree on all its connected ports.  Here I rely on the
initial HelloMsg to count the neighbors.  A cell assumes it has heard
from all its connected neighbors when the number of DiscoverD messages
is at least as big as the number of received Hello messages.  In this
version, I’m just verifying that I can get the counts correct.

This approach will fail if a neighbor sends a Hello message after the
cell has sent a DiscoverD for some tree.  However, there is still no
harm as long as no StackTree messages have reached that cell.  If they
have, the stacked trees won’t be extended to the next cell, but the
base tree will.

The key idea is that we only extend stacked trees by declaring a new
epoch, which involves flushing all application messages, extending the
stacked trees as appropriate, and then resuming application messages.