No more race condition
The code now does base tree building leaf to root using a heuristic to
decide if a cell is a leaf on a given tree.  If the number of received
DiscoverD messages is at least as large as the number of received Hello
messages, the cell assumes it has heard from all cells that received a
Discover from it on that tree.  Hence, if all the DiscoverD messages
are Subsequent, the cell is a leaf on that tree.  This heuristic should
be correct almost all the time.

StackTree and application messages obviously can be sent only after the
Discover process ends.  Cells can join later, but any previously
received StackTree or application messages will not be forwarded to
them automatically.  Nor does this version inform applications that a
new cell has joined.

THIS VERSION ASSUMES THE NOC CONNECTS TO A BORDER PORT BEFORE DISCOVER
FINISHES.