Added discover quiesce factor
My heuristic for knowing that discover had quiesced was to wait until a
cell had gotten as many DiscoverD messages for a given tree as it had
neighbors.  That often worked, but not always.  I’ve now added a
multiplier for the number of DiscoverD messages.  For the 10 cell
configuration, a factor of 4 appears to work.