Fixed a bug in the trace records
The thread_id in the trace header was the id of the spawning thread,
not that of the thread producing the trace record.  In fixing the bug,
I added a spawning_thread_id to the header.  The trace header now has
both the thread_id of the thread producing the record and the thread_id
of the thread that spawned it.