Noc no longer talks to cell agent via packet engine
The Noc was sending application messages from a border port through the
packet engine to the cell agent.  That caused an unnecessary
complication for the packet engine.  I am now connecting the border
port directly to the cell agent.

In order to get the traph processing in the right thread, the cell
agent bounces Noc messages off the cmodel.  I need to do the same thing
for the VM interface.  Right now it’s only working because the traphs
are fully set up before any VMs start.