Fixed some issues with previous version
There were a couple of problems with the previous version.  Mainly, the
cell agent to VM communication was only working by accident.  Since the
cell agent has different threads for talking to the cmodel and the VMs,
the copy of the data structures held in the VM thread wouldn’t update
when a change came in through the cmodel.  Hello World worked only
because there were no updates after the VMs started.

I fixed that problem with a kludge.  Application messages coming in
from the Noc via a border port or from a VM get bounced off the cmodel.
 The result is that the messages get processed by the thread that has
the correct state.

I was also sending the same TreeNameMsg to the Noc every time the root
of the tree received a StackTreeD.  That’s because I change from
stacking trees leaf ward to stacking them root ward.  The result is
that it’s inconvenient to check for duplicates in the cell agent.

Rather than fixing the problem in the cell agent, I use a kludge in
which the Noc tests if it has seen this AllowedTree before.