Some minor clean up
Removed “pub” from some functions that don’t need to be public.

Restored trace records to show hex rather than strings.

Found a Rustier way to convert between strings and arrays in name.rs.

The bin folder is now populated with the same files as in source.  I
didn’t do it, but it happened anyway.