More small changes
Changed TreeID from “C:1” to “Tree:C:1” to make it easier to
distinguish from CellID in the trace records.

Change the name of the unique message in the packet header to
unique_msg_id.  It was sender_msg_seq_no even though it was created
with rand().

Fixed the Display of StackTreeDPayload to include all fields.

I was converting messages to strings in the trace records, which made
parsing hard.  I now just use the object.