Use a single Cargo.toml and a single build.rs.
We had set up a separate folder and Cargo.toml for each executable,
sharing source code with symbolic links.  Rust has a different
approach, which I implemented.

All the files with a main() go into src/bin.  Cargo.toml gets a
separate [[bin]] section for each executable.  There is also a
[features] section that list the features that can be used in [#cfg()].

Which executable runs with which features is controlled on the command
line.  So, to run executable foo with feature bar, the command line is
cargo run —bin foo —features=“bar”.