Got cell.rs compiling with join handle
The join() method of join handle takes self as an argument, not &self,
which is necessary to prevent use of the handle following a join, but
it means the join handle can’t be part of a struct.  Hence, it’s not
possible for cell.rs to call a method in nalcell.rs to do the join.

The alternative is to return the join handle(s) to cell.rs and have it
do the join.  A minor annoyance is that the join handle also gets
returned to rack.rs, which doesn’t need it.

This version returns the packet engine join handle on the theory that a
cell that loses its cell agent and/or cmodel threads can still forward
packets.  However, if any of the three threads fails, we should
probably restart the cell.