Use ArrayString for name.rs
It’s really annoying having to clone IDs, so I made the name field a
fixed length array.  That change required conversions between the char
array and strings for printout and custom serialize/deserialize methods
to work with the trace record tools.

The arrayvec crate has an ArrayString struct that takes care of that
for me.  The result is a negative code day and, probably, better
performance.