Added AppDeleteTreeMsg and DeleteTreeMsg
I added the ability for the root of a tree to delete any tree stacked
on it.  There is now a new feature in Cargo.toml, delete_tree, that
tells the Master in service.rs to delete the AgentMaster tree.  I also
remove that tree from the allowed trees of all up trees to avoid the
reused tree name problem.

This version also replaces my custom code in name.rs that converts
between strings and fixed length arrays in order to make IDs implement
Copy.  I now use the arrayvec crate that does the same thing more
efficiently.