Reset parent of trees stacked on deleted tree
There are two options when deleting a stacked tree.  One is to delete
all trees stacked on it; the other, delete only the specified tree.
When doing the latter, trees stacked on the deleted tree will have the
deleted tree as a parent unless the parent is reset to be the parent of
the deleted tree.