Generate topology from HelloMsg
The HelloMsg tells each cell the sending cell_id and sender’s port
number for each HelloMsg received on a port.  This information can be
used to draw the connectivity diagram and to determine the forwarder of
every received message.

Both the geometry and topology are available to GET from localhost:8088.