Can now draw black trees
I’m drawing links on trees without arrow heads, because I don’t want to
have to figure out how to change the direction.

I found a problem with the tooltips that shows up if I try to show the
topology a second time.  For now I just reload the page.