Show small stacked tree
I can now show a stacked tree that does not include all cells in the
datacenter.  My bug was in specifying the GVM.  If XTND == true, then I
must add the cell to the tree so packets can be forwarded even if the
UP bit is not set.  In other words, I need a better test for tree
membership than just using SEND and RECV.