Add a class for stacked tree buttons
Eventually, each button will have its own color, but for now they’re
all the same.

I also fixed a problem where the project rebuilds when I change the
config file.  The problem is that cargo doesn’t know what files
build.rs might read, so it builds when any file in the project is
changed.  I added a special println! to build.rs to tell it to build
only if something in cellagent/src changes.  That works when running
cargo from the console but not from IntelliJ.