Got rid of stale buttons in the display
The problem was two-fold.  First, every time I run the simulator, I
send trace records to the server, which adds the relevant state.
However, it wasn’t starting from empty state on each run.  I now send a
trace record with a format of “reset” from datacenter.rs, which results
in a POST to that path.  At that path, I clear the state from a
previous run.

The second problem was that I wasn’t clearing the stacked tree buttons
from previous runs of the server.  The result was buttons for each
cells the NOC connected to.  The solution is to explicitly remove the
stacked tree buttons at the start of the visualize() function.