Got rid of redundant Discover messages
I was sending 3 times as many Discover messages for each root tree as I
should have.  The first mistake was sending the initial discover
message in both port_connected() and in process_hello_msg() when only
the latter was correct.  The second mistake was not quenching Discover
messages referring to the cell’s own tree, since those are sent in
process_hello_msg().