Can now view a smaller datacenter after viewing a bigger one
I was not clearing the server data structures when reading trace
records from a file.  The result was incorrect output or an error.

I also removed a double deserialization.  I was converting the Body as
Value to a string and then deserializing to the Body appropriate for
the trace record format.  Now, I’m going directly from Body as Value to
the appropriate Body.