Added filtered trace files for demo
Trace files are big, and most records are not relevant for displaying
the datacenter.  Replaying from the full file can be very slow.  These
two files contain only records relevant to the display.  Replay from
these files isn’t exactly fast, but it’s not nearly as painful as using
the unfiltered files.