Created a shell script for filtering trace files
./filter_traces.sh full_trace filtered_trace

where full_trace is the name of the trace file produced by the
emulator, and filtered_trace is the name of the file to contain the
filtered records.

The filtered file contains all the trace records having a format listed
in replay_from_file() in replay.rs.  Update the shell script if this
list changes.

I should really use an environment variable for the list of trace
record formats, but I don’t expect the list to change very often.