Add feature for conditionally using a web server
It is possible to send trace records to a web server and visualize them
in a browser.  The server URL is specified in an environment variable
SERVER_URL.  This commit introduces a feature “webserver” in Cargo.toml
of multicell/cellagent that makes use of the web server optional.

The README has instructions for using the web server.