Added to a debug trace record
I’m trying to find a heuristic for knowing when the discover process is
over.  I thought that I could do that if I knew how many trees a cell
had seen, but it didn’t work.  I’m leaving the “seen trees” field in
the debug trace record anyway.