Discovered a Heisenbug in black tree building
This version includes changes that reduce the frequency, but the bug
still pops up on one or two trees tree every few runs.  The cause is
lost DiscoverD NonParent messages.  I can see where they are sent on
all but one port in the trace, but some of the intended recipients
don’t show a trace record of having received it.

I don’t know where the problem lies.  I tried it on a 10-cell example,
and it never happens.  I tried turning on PE tracing on the 24-cell
model, and it never happens.  I have no idea how to find the cause of
the lost messages if I can’t get a trace of the PE when the failure
occurs.