Added ability to specify trace records to visualize in config file
Currently the visualizer needs trace records for dc and ca.  I had hard
wired that into config.rs, but then I need to rebuild if that set
changes.  In this version, I specify the set in the config file.

I also added a debug trace record in send_msg of cellagent.rs.