Changed visualization of black trees
I was showing the black trees root to leaf using the child ports, which
works fine unless there’s an error in the tree building.  Then it stops
prematurely, making it impossible to see which cell has the error.
This version shows leaf to root using the parent ports.  It’s now much
easier to identify where the tree is broken.