Physical neighbors are logical neighbors
Previously, the discover process was completely stochastic allowing a
node to first hear about a physical neighbor from a multi-hop
DiscoverMsg.  This version fixes that by having physical neighbors hold
a 1-hop DiscoverMsg until the cell that sent it has heard from all its
physical neighbors.

There’s still a heuristic involved because cells on the boundary don’t
have all their physical ports connected and because of potential
cabling errors.  The current heuristic is to wait for either all
physical ports to be connected or until the cell has seen a
configurable number of DiscoverMsgs.