Eliminated a gazillion unnecessary messages
I was sending DiscoverMsgs from the root every time the cell received a
DiscoverDMsg because I didn’t put parens in a compound if-test.  I was
also sending DiscoverDMsgs every time I received a DiscoverDMsg because
I forgot to put in an if-test.  The result is a trace file consisting
of about 40 records instead of 2,800.  Yikes.

This commit also has the start of doing depth first tree building to a
specified depth, but the code is only debugged for depth = 1.