Works for breadth first discover depth = 0
The resulting trees look similar to breadth first discover depth
= 1.  That’s because I don’t send any DiscoverMsgs until
CONFIG.min_hello HelloMsgs have ben received or the cell has received
at least CONFIG.min_discover DiscoverMsgs.  There are, however,
a few places where a physical neighbor of the root is not a child
of the root.

If I did send a DiscoverMsg on each HelloMsg, I would break the
connectivity heuristic and would need to go back to the timeout.