Tree stacking working again
A commit or two ago, I broke tree stacking but didn’t realize it.  The
error showed up on some runs when a cell didn’t have permission to use
one of the stacked trees.  The problem turned out to be sending a
Nonparent DiscoverDMsg on all ports instead of all except the one the
DiscoverMsg was received on.  The result was cells sometimes sending a
Parent DiscoverDMsg before receiving a Parent DiscoverDMsg from all
their children.

I am still sending unnecessary DiscoverDMsgs, but that doesn’t appear
to be causing any problems.