Trees build completely even if heuristic fails
A cell with all physical ports connected can send its discover
messages.  Other cells need some sort of heuristic.  We say the
heuristic failed if a port gets connected after the heuristic triggered.

This version correctly builds all trees even in the presence of
multiple heuristic failures.  The only problem on such failures is that
all physical neighbors may not be logical neighbors, but that case was
not seen in a number of runs.

Debugging this case was hard because it looked like some messages were
being lost, but the error never occurred when tracing the packet
engine.  This commit includes changes to reduce the overhead of
producing those trace records by replacing the bytes with the global
message sequence number.  (This trick only works on the simulator.)
The change allowed me to capture a failed run with packet engine trace
records.  I could then see that the problem was a failure to send some
messages on the late connected port.